Add a script to create stackless source archives.
A pure hg export isn't enough, because the source contains
generated files. If a generated files is less recent than its source
after extraction from the archive, make tries to recreate the
generated file. To avoid this the script created an archive,
that contains the members with correct time stamps in the
correct order.