Stackless issue #97: Fix stack switching for optimized builds
This change removes the "static" declaration from the function slp_switch().
This forces the compiler to adhere to the ABI specification.

On i386 and amd64 it is now save to configure stackless with
--enable-stacklessfewerregisters, except if you compile for darwin. The flag
disables explicit saving of %ebp/%rbp.

On Unix-like systems the source file slp_transfer.c now gets compiled
with the additional flag -fno-inline-functions instead of -O2.

https://bitbucket.org/stackless-dev/stackless/issues/98