Issue #121: Fix a reference leak, if Python runs Python code while clearing a thread state
If Python clears a thread state and a destructor or a weakref-callback runs Python code, Python used to create a new initial stub without clearing it later. This change reorders PyThreadState_Clear() to clear the PyStacklessState sub-structure last. A new test case for this problem uses thread local storage to execute a __del__-method.

https://bitbucket.org/stackless-dev/stackless/issues/121